+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+## 2.0.0
+- **Breaking**: Removed `Path.list()`
+- Added `Path.home()` constructor to get home directory
+- Added `Path.dir`, `Path.file`, and constructors `Path.fromFile()` & `Path.fromDir()` for increased interoperability with native `dart:io` `Directory` & `File` APIs
+- Documentation improvements
+
 ## 1.1.0
-- Added `iter()` method to list files in a directory
+- Added `list()` method to list files in a directory
 - Added `hashCode` override
 - Documentation improvements